+++ Form/Type/ProductType.php
@@ -49,9 +49,6 @@ class ProductType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('sku', 'text', array(
-                'label' => 'sylius_assortment.label.product.sku'
-            ))
             ->add('name', 'text', array(
                 'label' => 'sylius_assortment.label.product.name'
             ))