+++ Form/Type/VariantType.php
@@ -54,9 +54,6 @@ class VariantType extends AbstractType
                 'required' => false,
                 'label'    => 'sylius_assortment.label.variant.presentation'
             ))
-            ->add('sku', 'text', array(
-                'label' => 'sylius_assortment.label.variant.sku'
-            ))
             ->add('availableOn', 'datetime', array(
                 'label' => 'sylius_assortment.label.variant.available_on'
             ))